{"schema":"decodex/commit/1","summary":"Extract capture overlay controller","authority":"manual"}
